% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apacor.R
\name{apa.cor}
\alias{apa.cor}
\title{APA-Style Correlation Matrix}
\usage{
apa.cor(x, export = FALSE)
}
\arguments{
\item{x}{A dataframe with the variables of interest.}

\item{export}{Copies the resulted dataframe into the clipboard if set to TRUE (default is FALSE).}
}
\value{
A formatted dataframe representing the APA-style correlation matrix.
}
\description{
This function calculates a correlation matrix in APA (American Psychological Association) style
and returns the result as a formatted dataframe. Optionally, it can export the dataframe to the clipboard.
}
\examples{
# Using the airquality dataset as an example
x <- airquality
apa_correlation_matrix <- apa.cor(x, export = FALSE)
print(apa_correlation_matrix)
}
