% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apacor.R
\name{apa.cor}
\alias{apa.cor}
\title{A correlation matrix in R, using APA style}
\usage{
apa.cor(x, export = FALSE)
}
\arguments{
\item{x}{A dataframe with the variables of interest.}

\item{export}{Copies the resulted dataframe into the clipboard.}
}
\value{
A dataframe
}
\description{
A correlation matrix in R, using APA style
}
\examples{
x <- airquality
apa.cor(x, export = ",")
}
